third_party/WebKit/LayoutTests/webaudio/StereoPanner/stereopannernode-panning-expected.txt

CONSOLE WARNING: line 125: StereoPanner.pan.value -1.0625 outside nominal range [-1, 1]; value will be clamped.
CONSOLE WARNING: line 125: StereoPanner.pan.value -1.04104 outside nominal range [-1, 1]; value will be clamped.
CONSOLE WARNING: line 125: StereoPanner.pan.value -1.01957 outside nominal range [-1, 1]; value will be clamped.
CONSOLE WARNING: line 125: StereoPanner.pan.value 1.01957 outside nominal range [-1, 1]; value will be clamped.
CONSOLE WARNING: line 125: StereoPanner.pan.value 1.04104 outside nominal range [-1, 1]; value will be clamped.
CONSOLE WARNING: line 125: StereoPanner.pan.value 1.0625 outside nominal range [-1, 1]; value will be clamped.
CONSOLE WARNING: line 125: StereoPanner.pan.value -1.0625 outside nominal range [-1, 1]; value will be clamped.
CONSOLE WARNING: line 125: StereoPanner.pan.value -1.04104 outside nominal range [-1, 1]; value will be clamped.
CONSOLE WARNING: line 125: StereoPanner.pan.value -1.01957 outside nominal range [-1, 1]; value will be clamped.
CONSOLE WARNING: line 125: StereoPanner.pan.value 1.01957 outside nominal range [-1, 1]; value will be clamped.
CONSOLE WARNING: line 125: StereoPanner.pan.value 1.04104 outside nominal range [-1, 1]; value will be clamped.
CONSOLE WARNING: line 125: StereoPanner.pan.value 1.0625 outside nominal range [-1, 1]; value will be clamped.
Test panning model of StereoPannerNode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Number of impulses matches the number of panner nodes.
PASS All impulses at expected offsets.
PASS Left channel gain values are correct.
PASS Right channel gain values are correct.
PASS Test for mono input: passed.
PASS Number of impulses matches the number of panner nodes.
PASS All impulses at expected offsets.
PASS Left channel gain values are correct.
PASS Right channel gain values are correct.
PASS Test for stereo input: passed.
PASS successfullyParsed is true

TEST COMPLETE
